requirements.txt
appdirs==1.4.4
attrs==20.1.0
Babel==2.8.0
bcrypt==3.1.7
black==19.10b0
blinker==1.4
cffi==1.14.0
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
-e git://github.com/maxcountryman/flask-login.git@c073d3bdeceb2a9c280c421f57c9770ffd9f183a#egg=Flask_Login
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
identify==1.4.28
idna==2.9
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
nodeenv==1.5.0
-e git+https://github.com/onyxcherry/OnyxcherryOTP.git@1ab84ca79c4263ef228319d754d557843ac6b3c8#egg=OnyxcherryOTP
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.7.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyotp==2.3.0
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.13.0
python-http-client==3.3.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
sendgrid==6.4.6
six==1.15.0
SQLAlchemy==1.3.17
starkbank-ecdsa==1.0.0
toml==0.10.1
tox==3.19.0
typed-ast==1.4.1
virtualenv==20.0.31
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
